endpoints changed from app.py to route.py
